requirements.txt
aenum==2.0.2
argon2-cffi==16.2.0
asgi-redis==1.0.0
asgiref==1.0.0
autobahn==0.16.1
autopep8==1.2.4
boto==2.45.0
boto3==1.4.3
botocore==1.4.93
cffi==1.8.3
channels==0.17.3
constantly==15.1.0
daphne==0.15.0
defusedxml==0.4.1
Django==1.10.3
django-allauth==0.27.0
django-appconf==1.0.2
django-ipware==1.1.6
django-autocomplete-light==3.2.1
django-compressor==2.1
django-crispy-forms==1.6.0
django-endless-pagination==2.0
django-file-form==0.2.0
django-sass-processor==0.4.6
django-smtp-ssl==1.0
django-storages==1.5.2
djangorestframework==3.4.7
docutils==0.13.1
enum34==1.1.6
incremental==16.10.1
jmespath==0.9.2
libsass==0.11.1
msgpack-python==0.4.8
oauthlib==2.0.0
pathlib==1.0.1
pep8==1.7.0
Pillow==3.3.1
psycopg2==2.6.2
pycodestyle==2.2.0
pycparser==2.14
python-dateutil==2.5.3
python-memcached==1.58
python3-openid==3.0.10
rcssmin==1.0.6
redis==2.10.5
requests==2.11.1
requests-oauthlib==0.6.2
rjsmin==1.0.12
s3transfer==0.1.10
six==1.10.0
Twisted==16.5.0
txaio==2.5.2
zope.interface==4.3.2
